+++ pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
     "pyjwt>=2.10.1",
     "python-jose>=3.4.0",
     "python-multipart>=0.0.20",
+    "scipy>=1.15.3",
     "sqlalchemy>=2.0.41",
     "uvicorn>=0.34.2",
     "yfinance>=0.2.61",
